+++ pkg/kube/krt/helpers.go
@@ -17,28 +17,15 @@ package krt
 import (
 	"fmt"
 	"reflect"
-	"strings"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	acmetav1 "k8s.io/client-go/applyconfigurations/meta/v1"
-	"k8s.io/client-go/tools/cache"
 
-	"istio.io/api/type/v1beta1"
-	"istio.io/istio/pkg/config"
-	"istio.io/istio/pkg/kube/controllers"
 	"istio.io/istio/pkg/ptr"
 )
 
-func Filter[T any](data []T, f func(T) bool) []T {
-	fltd := make([]T, 0, len(data))
-	for _, e := range data {
-		if f(e) {
-			fltd = append(fltd, e)
-		}
-	}
-	return fltd
-}
-
+// GetKey returns the key for the provided object.
+// If there is none, this will panic.
 func GetKey[O any](a O) Key[O] {
 	if k, ok := tryGetKey[O](a); ok {
 		return k
@@ -51,48 +38,14 @@ func GetKey[O any](a O) Key[O] {
 	return ""
 }
 
-func tryGetKey[O any](a O) (Key[O], bool) {
-	as, ok := any(a).(string)
-	if ok {
-		return Key[O](as), true
-	}
-	ao, ok := any(a).(controllers.Object)
-	if ok {
-		k, _ := cache.MetaNamespaceKeyFunc(ao)
-		return Key[O](k), true
-	}
-	ac, ok := any(a).(config.Config)
-	if ok {
-		return Key[O](KeyFunc(ac.Name, ac.Namespace)), true
-	}
-	arn, ok := any(a).(resourceNamer)
-	if ok {
-		return Key[O](arn.ResourceName()), true
-	}
-	ack := GetApplyConfigKey(a)
-	if ack != nil {
-		return *ack, true
-	}
-	return "", false
-}
-
-func AppendNonNil[T any](data []T, i *T) []T {
-	if i != nil {
-		data = append(data, *i)
-	}
-	return data
-}
-
-func HasName(a any) bool {
-	_, ok := a.(controllers.Object)
-	return ok
-}
-
+// Named is a convenience struct. It is ideal to be embedded into a type that has a name and namespace,
+// and will automatically implement the various interfaces to return the name, namespace, and a key based on these two.
 type Named struct {
 	Name, Namespace string
 }
 
-func NewNamed(o metav1.ObjectMeta) Named {
+// NewNamed builds a Named object from a Kubernetes object type.
+func NewNamed(o metav1.Object) Named {
 	return Named{Name: o.GetName(), Namespace: o.GetNamespace()}
 }
 
@@ -106,58 +59,6 @@ func (n Named) GetName() string {
 
 func (n Named) GetNamespace() string {
 	return n.Namespace
-}
-
-type selector interface {
-	GetLabelSelector() map[string]string
-}
-
-type namer interface {
-	GetName() string
-}
-
-type namespacer interface {
-	GetNamespace() string
-}
-
-type labeler interface {
-	GetLabels() map[string]string
-}
-
-func GetName(a any) string {
-	ak, ok := a.(namer)
-	if ok {
-		return ak.GetName()
-	}
-	log.Debugf("No Name, got %T", a)
-	panic(fmt.Sprintf("No Name, got %T %+v", a, a))
-	return ""
-}
-
-func GetNamespace(a any) string {
-	ak, ok := a.(namespacer)
-	if ok {
-		return ak.GetNamespace()
-	}
-	panic(fmt.Sprintf("No Namespace, got %T", a))
-	return ""
-}
-
-func GetLabels(a any) map[string]string {
-	al, ok := a.(labeler)
-	if ok {
-		return al.GetLabels()
-	}
-	ak, ok := a.(metav1.Object)
-	if ok {
-		return ak.GetLabels()
-	}
-	ac, ok := a.(config.Config)
-	if ok {
-		return ac.Labels
-	}
-	panic(fmt.Sprintf("No Labels, got %T", a))
-	return nil
 }
 
 func GetApplyConfigKey[O any](a O) *Key[O] {
@@ -181,75 +82,11 @@ func GetApplyConfigKey[O any](a O) *Key[O] {
 	return ptr.Of(Key[O](*meta.Name))
 }
 
-func GetLabelSelector(a any) map[string]string {
-	ak, ok := a.(selector)
-	if ok {
-		return ak.GetLabelSelector()
-	}
-	val := reflect.ValueOf(a)
-
-	if val.Kind() == reflect.Ptr {
-		val = val.Elem()
-	}
-
-	specField := val.FieldByName("Spec")
-	if !specField.IsValid() {
-		log.Debugf("obj %T has no Spec", a)
-		return nil
-	}
-
-	labelsField := specField.FieldByName("Selector")
-	if !labelsField.IsValid() {
-		log.Debugf("obj %T has no Selector", a)
-		return nil
-	}
-
-	switch s := labelsField.Interface().(type) {
-	case *v1beta1.WorkloadSelector:
-		return s.GetMatchLabels()
-	case map[string]string:
-		return s
-	default:
-		log.Debugf("obj %T has unknown Selector", s)
-		return nil
-	}
-}
-
-func GetType[T any]() reflect.Type {
-	t := reflect.TypeOf(*new(T))
-	if t.Kind() != reflect.Struct {
-		return t.Elem()
-	}
-	return t
-}
-
-func GetTypeOf(a any) reflect.Type {
-	t := reflect.TypeOf(a)
-	if t.Kind() != reflect.Struct {
-		return t.Elem()
-	}
-	return t
-}
-
-// KeyFunc is the internal API key function that returns "namespace"/"name" or
+// keyFunc is the internal API key function that returns "namespace"/"name" or
 // "name" if "namespace" is empty
-func KeyFunc(name, namespace string) string {
+func keyFunc(name, namespace string) string {
 	if len(namespace) == 0 {
 		return name
 	}
 	return namespace + "/" + name
-}
-
-func SplitKeyFunc(key string) (namespace, name string) {
-	parts := strings.Split(key, "/")
-	switch len(parts) {
-	case 1:
-		// name only, no namespace
-		return "", parts[0]
-	case 2:
-		// namespace and name
-		return parts[0], parts[1]
-	}
-
-	return "", ""
 }
