+++ pkg/kube/krt/testing.go
@@ -1,5 +1,18 @@
-package krt
+// Copyright Istio Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
+package krt
 
 type dump interface {
 	Dump()
@@ -13,4 +26,3 @@ func Dump[O any](c Collection[O]) {
 		log.Warnf("cannot dump collection %T", c)
 	}
 }
-
